refactor into rows
