+++ app/assets/stylesheets/components/list.scss
@@ -1,13 +1,3 @@
-.items {
-  display:inline-block;
-  padding: 10px;
-  vertical-align:top;
-  padding-top: 10px;
-  .li-text {
-    padding:4px;
-  }
-}
-
 #labels {
   font-weight: bold;
 }
@@ -26,6 +16,12 @@
     display: flex;
     flex-direction: row;
     align-items: center;
+
+    .img {
+      object-fit: cover;
+      width: 150px;
+      height: 150px;
+    }
   }
 }
 
@@ -78,7 +74,7 @@
 
     .attr-item {
       @extend .mb2;
-      
+
       &:last-child {
         margin: 0;
       }
