+++ app/controllers/api/artists_controller.rb
@@ -38,27 +38,8 @@ class Api::ArtistsController < ApplicationController
   def requests
     artist = Artist.find(params[:id])
     requests = artist.requests
-    requests_full = []
-    requests.each do |request|
-      this_request = {}
-      this_request["request"] = request
-      this_request["buyer"] = request.buyer
-      this_request["artist"] = artist
-      this_request["receipt"] = request.receipt
-      requests_full.append(this_request)
-    end
-    if requests
-      render json: requests_full.to_json
-    else
-      render_json_message(:forbidden, errors: requests.errors.full_messages)
-    end
-  end
-
-  def receipts
-    artist = Artist.find(params[:id])
-    receipts = artist.receipts
-    render json: receipts,
-        each_serializer: ReceiptSerializer
+    render json: requests,
+        each_serializer: RequestSerializer
   end
 
 end
