+++ app/javascript/components/artists/AllArtists.jsx
