+++ app/javascript/components/artists/ArtistRow.jsx
