+++ app/serializers/request_serializer.rb
@@ -1,0 +1,24 @@
+class RequestSerializer < ActiveModel::Serializer
+  attributes :id, :message, :open, :created_at, :updated_at, :types, :artist, :buyer, :work, :receipt
+
+  belongs_to :work
+  belongs_to :buyer
+  belongs_to :artist
+  has_one :receipt
+
+  def work
+    object.artist
+  end
+
+  def buyer
+    object.buyer
+  end
+
+  def work
+    object.work
+  end
+
+  def receipt
+    object.receipt
+  end
+end
