+++ app/views/artists/all_artists.html.slim
@@ -1,2 +1,2 @@
 .container.dashboard-container
-  = react_component "AllArtists", {}
+  = react_component "artists/AllArtists", {}
