+++ app/views/artists/show.html.slim
@@ -1,2 +1,2 @@
 .container.dashboard-container
-  = react_component "ArtistProfile", {artist: @artist, buyer: current_buyer}
+  = react_component "artists/ArtistProfile", {artist: @artist, buyer: current_buyer}
