--- app/views/artists/transactions.html.slim
@@ -1,2 +1,0 @@
-.container.dashboard-container
-  = react_component "ArtistTransactions", {artist: current_artist}
