+++ app/views/buyers/show.html.slim
@@ -1,2 +1,2 @@
 .container.dashboard-container
-  = react_component "BuyerProfile", {buyer: @buyer}
+  = react_component "buyers/BuyerProfile", {buyer: @buyer}
