+++ app/views/receipts/new.html.slim
@@ -1,2 +1,2 @@
 .container.dashboard-container
-  = react_component "CreateTransaction", { artist: current_artist }
+  = react_component "transactions/CreateTransaction", { artist: current_artist }
