+++ app/views/requests/home.html.slim
@@ -1,2 +1,2 @@
 .container.dashboard-container
-  = react_component "ArtistRequests", {artist: @artist, buyer: @buyer}
+  = react_component "requests/UserRequests", {artist: @artist, buyer: @buyer}
