+++ app/views/works/index.html.slim
@@ -1,2 +1,2 @@
 .container.dashboard-container
-  = react_component "Works"
+  = react_component "works/Works"
