+++ app/views/works/new.html.slim
@@ -1,2 +1,2 @@
 .container.dashboard-container
-  = react_component "CreateWork", {artist: current_artist}
+  = react_component "works/CreateWork", {artist: current_artist}
