+++ app/views/works/show.html.slim
@@ -1,2 +1,2 @@
 .container.dashboard-container
-  = react_component "DetailedWork", {work: @work, buyer: current_buyer}
+  = react_component "works/DetailedWork", {work: @work, buyer: current_buyer}
